requirements.txt
black==23.9.1
functions-framework==3.*
google-cloud-bigquery==3.19.0
google-cloud-storage==2.9.0

